+++ test/tests.dart
@@ -16,8 +16,8 @@ void main() {
     // Outputs current summoner level
     // etc.
     league.getGameHistory(accountID: summonerInfo.accID).then((games){
-      games[5].stats().then((stats){
-        print(stats.win);
+      games[0].stats().then((stats){
+        print(stats.kills);
       });
     });
   });