+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.0.1] - Initial Release :D
 
 ## [0.0.2] - Now: WIN option available! -> getGameHistory
+
+## [0.0.3] - Now: Kills/Deaths/Assists option available! -> getGameHistory
+
+## [0.0.4] - Now: CS option available! -> getGameHistory